move DocstringStyle enum to darglint.docstring.style module
