+++ tests/test_integrity_checker.py
@@ -5,9 +5,7 @@ from unittest import (
 )
 
 from darglint.strictness import Strictness
-from darglint.docstring.base import (
-    DocstringStyle,
-)
+from darglint.docstring.style import DocstringStyle
 from darglint.integrity_checker import (
     IntegrityChecker,
 )
