+++ tests/test_numpy_parser.py
@@ -4,9 +4,7 @@ from unittest import (
     TestCase,
     skip,
 )
-from darglint.docstring.base import (
-    DocstringStyle,
-)
+from darglint.docstring.style import DocstringStyle
 from darglint.strictness import Strictness
 from darglint.lex import (
     condense,
